Don't use sensor collection for velocity data
This is a bug leftover from a while back. It's very subtle so it took a while to catch, but basically sensor phase/motor inversions are not reflected in the sensor collection data; that's just raw encoder data. Therefore, this data will be incorrect in the case of sensor phase/inversion.

The use-case of explicitly checking the feedback device is no longer valid, since nowhere in the library do we mess with feedback device anymore.