Added new swerve path following algorithms
These have been tested using the wpilib simulations. They should be tested on a robot during the offseason.

The FrcSwervePathFollower is basically a wrapper/converter around wpilibs trajectory follower for swerve, and there are some utilities to help with the creation of those paths. These paths can still be generated at runtime.